+++ src/contexts/index.tsx
@@ -12,3 +12,4 @@ export { SelectedWordsContext } from "./phrase-bubble";
 export { FdaVersionsContext, FdaVersionsProvider } from "./fda-versions";
 export { SearchSupportContext } from "./search-support";
 export { TableSelectContext, TableSelectProvider } from "./table-select";
+export { LocalGenericContext, LocalGenericProvider } from "./generic-local";
