+++ src/http/internal/index.ts
@@ -13,3 +13,4 @@ export async function fetchApiRoot(token: string) {
 export * from "./logout";
 export * from "./aws";
 export * from "./db";
+export * from "./cookie";
