+++ src/middleware.ts
@@ -5,7 +5,6 @@ import { redirect } from "next/navigation";
 import { cookies } from "next/headers";
 
 const openid_conf_uri = process.env.NEXT_PUBLIC_COGNITO_OPENID_CONF_URI;
-console.log(openid_conf_uri);
 
 export async function middleware(request: NextRequest) {
   const cookie = cookies();
